Factorize e2e tests
This commit provides a better reuse of common setup functions by
extending our e2e actions. This allows shortening the tests and improve
clarity.
