+++ tests/e2e/e2e_actions_test.go
@@ -26,6 +26,7 @@ import (
 	"github.com/NVIDIA/dcgm-exporter/tests/e2e/internal/framework"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/rest"
 	restclient "k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -76,6 +77,38 @@ func shouldCreateHelmClient(config *rest.Config) *framework.HelmClient {
 	return helmClient
 }
 
+func shouldInstallHelmChart(ctx context.Context, helmClient *framework.HelmClient, values []string) string {
+	_, _ = fmt.Fprintf(GinkgoWriter, "Helm chart installation: %q chart started.\n",
+		testContext.chart)
+
+	if testContext.arguments != "" {
+		values = append(values, fmt.Sprintf("arguments=%s", testContext.arguments))
+	}
+
+	if testContext.imageRepository != "" {
+		values = append(values, fmt.Sprintf("image.repository=%s", testContext.imageRepository))
+	}
+	if testContext.imageTag != "" {
+		values = append(values, fmt.Sprintf("image.tag=%s", testContext.imageTag))
+	}
+
+	helmReleaseName, err := helmClient.Install(ctx, values, framework.HelmChartOptions{
+		CleanupOnFail: true,
+		GenerateName:  true,
+		Timeout:       5 * time.Minute,
+		Wait:          true,
+		DryRun:        false,
+	})
+	Expect(err).ShouldNot(HaveOccurred(), "Helm chart installation: %q chart failed with error err: %v", testContext.chart, err)
+
+	_, _ = fmt.Fprintf(GinkgoWriter, "Helm chart installation: %q completed.\n",
+		testContext.chart)
+	_, _ = fmt.Fprintf(GinkgoWriter, "Helm chart installation: new %q release name.\n",
+		helmReleaseName)
+
+	return helmReleaseName
+}
+
 func shouldUninstallHelmChart(helmClient *framework.HelmClient, helmReleaseName string) {
 	if helmClient != nil && helmReleaseName != "" {
 		_, _ = fmt.Fprintf(GinkgoWriter, "Helm chart uninstall: release %q of the helm chart: %q started.\n",
@@ -122,4 +155,118 @@ func shouldDeleteNamespace(ctx context.Context, kubeClient *framework.KubeClient
 
 		_, _ = fmt.Fprintf(GinkgoWriter, "Namespace deletion: %q namespace fully deleted.\n", testContext.namespace)
 	}
+}
+
+func shouldCreateDCGMPod(ctx context.Context, kubeClient *framework.KubeClient, namespace string, labels map[string]string) *corev1.Pod {
+	_, _ = fmt.Fprintln(GinkgoWriter, "Pod creation verification: started")
+
+	var dcgmExpPod *corev1.Pod
+
+	Eventually(func(ctx context.Context) bool {
+		pods, err := kubeClient.GetPodsByLabel(ctx, testContext.namespace, labels)
+		if err != nil {
+			Fail(fmt.Sprintf("Pod creation: Failed with error: %v", err))
+			return false
+		}
+
+		if len(pods) == 1 {
+			dcgmExpPod = &pods[0]
+			return true
+		}
+
+		return false
+	}).WithPolling(time.Second).Within(15 * time.Minute).WithContext(ctx).Should(BeTrue())
+
+	_, _ = fmt.Fprintln(GinkgoWriter, "Pod creation verification: completed")
+
+	return dcgmExpPod
+}
+
+func shouldCreateWorkloadPod(ctx context.Context, kubeClient *framework.KubeClient, labels map[string]string) {
+	_, _ = fmt.Fprintln(GinkgoWriter, "Workload pod creation: started")
+
+	workloadPod, err := kubeClient.CreatePod(ctx,
+		testContext.namespace,
+		labels,
+		workloadPodName,
+		workloadContainerName,
+		workloadImage,
+	)
+
+	Expect(err).ShouldNot(HaveOccurred(),
+		"Workload pod creation: Failed create workload pod with err: %v", err)
+	Eventually(func(ctx context.Context) bool {
+		isReady, err := kubeClient.CheckPodStatus(ctx,
+			testContext.namespace,
+			workloadPod.Name, func(namespace, podName string, status corev1.PodStatus) (bool, error) {
+				return status.Phase == corev1.PodSucceeded, nil
+			})
+		if err != nil {
+			Fail(fmt.Sprintf("Workload pod creation: Checking pod status: Failed with error: %v", err))
+		}
+
+		return isReady
+	}).WithPolling(time.Second).Within(15 * time.Minute).WithContext(ctx).Should(BeTrue())
+
+	_, _ = fmt.Fprintln(GinkgoWriter, "Workload pod creation: completed")
+}
+
+func shouldEnsurePodReadiness(ctx context.Context, kubeClient *framework.KubeClient, dcgmExpPod *corev1.Pod) {
+	_, _ = fmt.Fprintln(GinkgoWriter, "Checking pod status: started")
+
+	Eventually(func(ctx context.Context) bool {
+		isReady, err := kubeClient.CheckPodStatus(ctx,
+			testContext.namespace,
+			dcgmExpPod.Name,
+			func(namespace, podName string, status corev1.PodStatus) (bool, error) {
+				for _, c := range status.Conditions {
+					if c.Type != corev1.PodReady {
+						continue
+					}
+					if c.Status == corev1.ConditionTrue {
+						return true, nil
+					}
+				}
+
+				for _, c := range status.ContainerStatuses {
+					if c.State.Waiting != nil && c.State.Waiting.Reason == "CrashLoopBackOff" {
+						return false, fmt.Errorf("pod %s in namespace %s is in CrashLoopBackOff", podName, namespace)
+					}
+				}
+
+				return false, nil
+			})
+		if err != nil {
+			Fail(fmt.Sprintf("Checking pod status: Failed with error: %v", err))
+		}
+
+		return isReady
+	}).WithPolling(time.Second).Within(15 * time.Minute).WithContext(ctx).Should(BeTrue())
+
+	_, _ = fmt.Fprintln(GinkgoWriter, "Checking pod status: completed")
+}
+
+func shouldReadMetrics(ctx context.Context, kubeClient *framework.KubeClient, dcgmExpPod *corev1.Pod, dcgmExporterPort uint) []byte {
+	_, _ = fmt.Fprintln(GinkgoWriter, "Read metrics: started")
+
+	var metricsResponse []byte
+
+	Eventually(func(ctx context.Context) bool {
+		var err error
+
+		metricsResponse, err = kubeClient.DoHttpRequest(ctx,
+			testContext.namespace,
+			dcgmExpPod.Name,
+			dcgmExporterPort,
+			"metrics")
+		if err != nil {
+			Fail(fmt.Sprintf("Read metrics: Failed with error: %v", err))
+		}
+
+		return len(metricsResponse) > 0
+	}).WithPolling(time.Second).Within(time.Minute).WithContext(ctx).Should(BeTrue())
+
+	_, _ = fmt.Fprintln(GinkgoWriter, "Read metrics: completed")
+
+	return metricsResponse
 }
